removed (this.state.value) from form inputs
